backend/requirements.txt
Django>=4.2
djangorestframework
mysqlclient
pandas
numpy
scipy
statsmodels
matplotlib
python-magic
python-dateutil
django-cors-headers